+++ docs/src/routes/+layout.svelte
@@ -1,7 +1,23 @@
-<header>
-  SvelementUI
-  <nav>
-    <a href="/component">Component</a>
-  </nav>
-</header>
-<slot/>
+<SvelContainer>
+  <SvelHeader>
+    SvelementUI header
+    <nav>
+      <a href="{PUBLIC_BASE_PATH}/component">Component</a>
+    </nav>
+  </SvelHeader>
+
+  <SvelMain>
+    <slot/>
+  </SvelMain>
+  <SvelFooter>
+    SvelementUI footer
+  </SvelFooter>
+</SvelContainer>
+
+<script lang="ts">
+import { PUBLIC_BASE_PATH } from '$env/static/public';
+import SvelContainer from '@svelement-ui/container';
+import SvelHeader from '@svelement-ui/header';
+import SvelMain from '@svelement-ui/main';
+import SvelFooter from '@svelement-ui/footer';
+</script>