+++ docs/src/routes/+page.svelte
@@ -1,2 +1,2 @@
-<h1>Welcome to SvelteKit</h1>
-<p>Visit <a href="https://kit.svelte.dev">kit.svelte.dev</a> to read the documentation</p>
+<h1>Welcome to SvelementUI</h1>
+<p>Visit <a href="https://koory1st.github.io/svelement-ui">koory1st.github.io/svelement-ui</a> to read the documentation</p>