+++ packages/aside/src/lib/svel-aside.svelte
@@ -1,0 +1,7 @@
+<script lang="ts">
+  export let width = '300px'
+</script>
+
+<aside class="svel-aside" style={`width:${width}`}>
+  <slot />
+</aside>