+++ packages/header/src/lib/svel-header.svelte
@@ -1,0 +1,7 @@
+<script lang="ts">
+  export let height = '60px'
+</script>
+
+<header class="svel-header" style={`height:${height}`}>
+  <slot />
+</header>