+++ packages/main/src/lib/index.ts
@@ -1,0 +1,2 @@
+import SvelMain from './svel-main.svelte'
+export default SvelMain