+++ packages/main/src/lib/svel-main.svelte
@@ -1,0 +1,3 @@
+<main class="svel-main">
+  <slot />
+</main>