+++ CHANGELOG.md
@@ -3,6 +3,43 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.6.0](https://github.com/koory1st/svelement-ui/compare/v0.5.0...v0.6.0) (2023-08-13)
+
+
+### Bug Fixes
+
+* **aside:** change class name ([1e8d0af](https://github.com/koory1st/svelement-ui/commit/1e8d0af50540ffab592009716ccf0e0c17f029b0))
+* **container:** change class name ([ae15cfc](https://github.com/koory1st/svelement-ui/commit/ae15cfc0e574f970f4804118b011c85cc379e9f6))
+* **docs:** add components ([b42ffe2](https://github.com/koory1st/svelement-ui/commit/b42ffe20905f95430387d7c4068b5c584dd4e500))
+
+
+### Features
+
+* **aside:** add ([87f772d](https://github.com/koory1st/svelement-ui/commit/87f772d9b78e7755358630e424f491cc34f9a844))
+* **footer:** add ([2c5e6b9](https://github.com/koory1st/svelement-ui/commit/2c5e6b990083aec3d33c6a9b6a41a17a9fd752ac))
+* **header:** add ([2c10f3c](https://github.com/koory1st/svelement-ui/commit/2c10f3c44deb176cc1991150ef9d58828701a8e6))
+* **main:** add ([0b4aa42](https://github.com/koory1st/svelement-ui/commit/0b4aa421e0fc116804236bb2e84cd1f7f7a7992f))
+
+
+
+
+
+# [0.5.0](https://github.com/koory1st/svelement-ui/compare/v0.4.3...v0.5.0) (2023-08-12)
+
+
+### Bug Fixes
+
+* **docs:** use absolute url ([285003c](https://github.com/koory1st/svelement-ui/commit/285003c12442917bf870004fbc4e5724a1d5b8ba))
+
+
+### Features
+
+* **container:** add ([eb6c447](https://github.com/koory1st/svelement-ui/commit/eb6c4478e6d1a7133946ecbc7d10c39ea84b40f9))
+
+
+
+
+
 ## [0.4.3](https://github.com/koory1st/svelement-ui/compare/v0.4.1...v0.4.3) (2023-08-11)