+++ packages/util-array-2-class-string/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.6.0](https://github.com/koory1st/svelement-ui/compare/v0.5.0...v0.6.0) (2023-08-13)
+
+**Note:** Version bump only for package @svelement-ui/util-array-2-class-string
+
+
+
+
+
+# [0.5.0](https://github.com/koory1st/svelement-ui/compare/v0.4.3...v0.5.0) (2023-08-12)
+
+**Note:** Version bump only for package @svelement-ui/util-array-2-class-string
+
+
+
+
+
 ## [0.4.3](https://github.com/koory1st/svelement-ui/compare/v0.4.1...v0.4.3) (2023-08-11)
 
 **Note:** Version bump only for package @svelement-ui/util-array-2-class-string